Fix stylus format
As stylus is white-space sensitive the indentation led to an error.
Also the assignment operator needs to be `=`.

We should also use the opportunity of the stylus format not being used widely yet (at least that is what I assume, as the error would have come up earlier) to add the leading `$` to variables. This is not necessary, but considered best practice – for details see https://gist.github.com/declandewet/7220997#variable-naming